pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stv_scraper"
version = "0.0.0+Dev.Bing.GoogleChrome"
authors = [
    { name = "星灿长风v(StarWindv)" },
]
description = "Bing图片爬虫工具"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.7"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "selenium>=4.0",
    "requests>=2.26",
    "Pillow>=9.0",
    "imagehash>=4.3",
]

[project.scripts]
stv_scraper = "stv_scraper.main:main"